This changes merges the functionality of the `cafe_tree.R` script and the `highlight_nodes.R` scripts into one.
Previously `cafe_tree.R` and `highlight_nodes.R` did basically the same thing but except that `cafe_tree.R` used the labels in the newikc and `highlight_nodes.R` used the labels as received from parsing.

Now the annotation table is optional, and `--label` flag can be used to determine which labels to use. This is boolean based `--label` is false by default and will use the parsed labels.

If the `--label` flag is used and given `TRUE` it will use the newick labels.